% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customiser-package.R
\docType{package}
\name{customiser-package}
\alias{customiser}
\alias{customiser-package}
\title{customiser: Customise your R experience}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: James Laird-Smith \email{jameslairdsmith@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
