% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customiser.R
\name{rmarkdown_simple}
\alias{rmarkdown_simple}
\alias{rmarkdown_no_overwrite}
\alias{rmarkdown_allow_overwrite}
\title{Example R markdown files}
\usage{
rmarkdown_simple()

rmarkdown_no_overwrite()

rmarkdown_allow_overwrite()
}
\value{
Character path to the example file.
}
\description{
Exclusively for testing and demonstration, the package comes with a few
simple R Markdown files.
}
\examples{
rmarkdown_simple()
rmarkdown_no_overwrite()
rmarkdown_allow_overwrite()
}
